EES-242 - initial refactor of Table Tool into common project
