+++ src/explore-education-statistics-common/src/modules/table-tool/components/FixedMultiHeaderDataTable.module.scss
@@ -1,4 +1,4 @@
-@import '~src/styles/govuk-base';
+@import '../../../styles/govuk-base';
 
 .figure {
   margin: 0;
