+++ src/explore-education-statistics-common/src/modules/table-tool/components/FormFieldSortableListGroup.module.scss
@@ -1,4 +1,4 @@
-@import '~src/styles/govuk-base';
+@import '../../../styles/govuk-base';
 
 .groupsFieldset {
   border: 4px solid govuk-colour('grey-3');
