+++ src/explore-education-statistics-common/src/modules/table-tool/components/FormSortableList.module.scss
@@ -1,4 +1,4 @@
-@import '~src/styles/govuk-base';
+@import '../../../styles/govuk-base';
 
 .listDraggingOver {
   background: govuk-colour('grey-4');
