+++ src/explore-education-statistics-common/src/modules/table-tool/components/GroupedDataTable.module.scss
@@ -1,4 +1,4 @@
-@import '~explore-education-statistics-common/src/styles/govuk-base';
+@import '../../../styles/govuk-base';
 
 .groupHeadingRow {
   background: govuk-colour('grey-4');
