+++ src/explore-education-statistics-common/src/modules/table-tool/components/MultiHeaderTable.module.scss
@@ -1,4 +1,4 @@
-@import '';
+@import '../../../styles/govuk-base';
 
 $border-width: 4px;
