+++ src/explore-education-statistics-common/src/modules/table-tool/components/TableHeadersForm.module.scss
@@ -1,4 +1,4 @@
-@import '~src/styles/govuk-base';
+@import '../../../styles/govuk-base';
 
 .groupsContainer {
   padding: govuk-spacing(2);
@@ -15,11 +15,11 @@
 }
 
 .isDraggingOver {
-  composes: isDraggingOver from 'FormFieldSortableListGroup.module.scss';
+  composes: isDraggingOver from './FormFieldSortableListGroup.module.scss';
 }
 
 .list {
-  composes: list from 'FormFieldSortableListGroup.module.scss';
+  composes: list from './FormFieldSortableListGroup.module.scss';
 
   &:not(:last-of-type) {
     margin-bottom: govuk-spacing(3);
@@ -27,5 +27,5 @@
 }
 
 .isDragging {
-  composes: isDragging from 'FormFieldSortableListGroup.module.scss';
+  composes: isDragging from './FormFieldSortableListGroup.module.scss';
 }
