+++ src/explore-education-statistics-common/src/modules/table-tool/components/WizardStep.module.scss
@@ -1,4 +1,4 @@
-@import '~src/styles/govuk-base';
+@import '../../../styles/govuk-base';
 
 $circle-size: 40px;
 $circle-border-size: 4px;
